Stock Aggregation- Batch process.
Stock Aggregation using Apache Spark - Batch process.

This is a project based on "Apache Spark". We want to build a pipeline that will do stock aggregation across multiple locations for an enterprise. This enterprise has warehouses across the globe. Warehouses maintain stock of items by location and distribute them to local stores. Each warehouse also has a local data centre with all the required hardware and software deployed in that centre. A stock management application runs in each warehouse. It is the same software, but has local independent instances deployed in the local data centre. A local MariaDB database keeps track of warehouse stock information. Stock is maintained by each item for each day. We keep track of the opening stock, receipts, and issues for each item for each day. The enterprise wants to create a consolidated central stock database. Item stock information needs to be collected from across warehouses on a daily basis, and then consolidated in a central MariaDB database. We need to set up batch processing to upload local warehouse data into a central data cloud, and then aggregate stock across locations. The solution should be scalable to hundreds of warehouses. This is a simple use case, but representative of a number of real life use cases that require data gathering, consolidation, and aggregation.